+++ src/datatype/mod.rs
@@ -10,5 +10,5 @@ pub use evaluate::evaluate_construct;
 #[cfg(feature = "typechecker")]
 pub(crate) use evaluate::shorthand_rebind;
 pub use lookup::StaticLookup;
-pub use types::{Datatype, variant_type_name};
+pub use types::{Datatype, variant_type_from_name, variant_type_name};
 pub use variants::EnumItemFromNameAndValueName;